requirements.txt
pytest==8.4.2
requests==2.32.5
pytest-rerunfailures==16.1
allure-pytest==2.15.0